Refactor Pie chart pt1
