+++ Sources/SwiftUICharts/BarChart/BarChartRow.swift
@@ -1,6 +1,6 @@
 //
 //  ChartRow.swift
-//  ChartView
+//  SwiftUICharts
 //
 //  Created by András Samu on 2019. 06. 12..
 //  Copyright © 2019. András Samu. All rights reserved.
