+++ Sources/SwiftUICharts/BarChart/LabelView.swift
@@ -1,6 +1,6 @@
 //
 //  LabelView.swift
-//  BarChart
+//  SwiftUICharts
 //
 //  Created by Samu András on 2020. 01. 08..
 //  Copyright © 2020. Samu András. All rights reserved.
