+++ Sources/SwiftUICharts/Common/ChartStyle.swift
@@ -1,8 +1,9 @@
 //
-//  File.swift
-//
+//  ChartStyle.swift
+//  SwiftUICharts
 //
 //  Created by Pawel on 09/08/2020.
+//  Copyright © 2020 Pawel Wiszenko. All rights reserved.
 //
 
 import SwiftUI
